keyring: Keyring (impl), not interface
